chore: reorganizado entidades do sistema
