+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/domain/models/cliente/Cartao.java
@@ -1,6 +1,7 @@
-package br.com.fatecmc.tecladostorelesbackend.domain.models;
+package br.com.fatecmc.tecladostorelesbackend.domain.models.cliente;
 
-import br.com.fatecmc.tecladostorelesbackend.domain.models.enums.cartoes.Bandeiras;
+import br.com.fatecmc.tecladostorelesbackend.domain.models.BaseModel;
+import br.com.fatecmc.tecladostorelesbackend.domain.models.cliente.enums.cartoes.Bandeiras;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @EqualsAndHashCode(of = "id")
 @Table(name = "cartoes")
-public class Cartao extends BaseModel{
+public class Cartao extends BaseModel {
 
     private Boolean preferencial;
     private String numeros;
