+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/domain/models/cliente/Cliente.java
@@ -1,5 +1,6 @@
-package br.com.fatecmc.tecladostorelesbackend.domain.models;
+package br.com.fatecmc.tecladostorelesbackend.domain.models.cliente;
 
+import br.com.fatecmc.tecladostorelesbackend.domain.models.BaseModel;
 import jakarta.persistence.*;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
