+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/domain/models/cliente/enums/cartoes/Bandeiras.java
@@ -1,0 +1,5 @@
+package br.com.fatecmc.tecladostorelesbackend.domain.models.cliente.enums.cartoes;
+
+public enum Bandeiras {
+    VISA, MASTERCARD
+}
