+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/domain/models/cliente/enums/enderecos/TiposEndereco.java
@@ -1,0 +1,5 @@
+package br.com.fatecmc.tecladostorelesbackend.domain.models.cliente.enums.enderecos;
+
+public enum TiposEndereco {
+    RESIDENCIAL, COBRANCA, ENTREGA
+}
