--- src/main/java/br/com/fatecmc/tecladostorelesbackend/domain/models/enums/cartoes/Bandeiras.java
@@ -1,5 +1,0 @@
-package br.com.fatecmc.tecladostorelesbackend.domain.models.enums.cartoes;
-
-public enum Bandeiras {
-    VISA, MASTERCARD
-}
