--- src/main/java/br/com/fatecmc/tecladostorelesbackend/domain/models/enums/enderecos/TiposEndereco.java
@@ -1,5 +1,0 @@
-package br.com.fatecmc.tecladostorelesbackend.domain.models.enums.enderecos;
-
-public enum TiposEndereco {
-    RESIDENCIAL, COBRANCA, ENTREGA
-}
